+++ README.md
@@ -93,3 +93,56 @@ DSL(Domain-specific language) ↔ 범용 프로그래밍 언어
 - 함수(또는 메서드)의 길이가 15라인을 넘어가지 않도록 구현한다.
 - 기능을 구현하기 전에 README.md 파일에 구현할 기능 목록을 정리해 추가한다.
 - git의 commit 단위는 앞 단계에서 README.md 파일에 정리한 기능 목록 단위로 추가한다.
+
+
+## STEP3 (3단계 - 블랙잭)
+### 기능 요구사항
+블랙잭 게임을 변형한 프로그램을 구현한다. 블랙잭 게임은 딜러와 플레이어 중 카드의 합이 21 또는 21에 가장 가까운 숫자를 가지는 쪽이 이기는 게임이다.
+
+- 카드의 숫자 계산은 카드 숫자를 기본으로 하며, 예외로 Ace는 1 또는 11로 계산할 수 있으며, King, Queen, Jack은 각각 10으로 계산한다.
+- 게임을 시작하면 플레이어는 두 장의 카드를 지급 받으며, 두 장의 카드 숫자를 합쳐 21을 초과하지 않으면서 21에 가깝게 만들면 이긴다. 21을 넘지 않을 경우 원한다면 얼마든지 카드를 계속 뽑을 수 있다.
+- 딜러는 처음에 받은 2장의 합계가 16이하이면 반드시 1장의 카드를 추가로 받아야 하고, 17점 이상이면 추가로 받을 수 없다.
+- 딜러가 21을 초과하면 그 시점까지 남아 있던 플레이어들은 가지고 있는 패에 상관 없이 승리한다.
+- 게임을 완료한 후 각 플레이어별로 승패를 출력한다.
+
+### 기능 목록
+### 기능 목록
+- [x] 이름은 1 ~ 10자 이내이여야 한다.
+  - 이름에 빈 문자열이 들어올 경우 IllegalArgumentException 예외를 던진다.
+  - 이름에 공백, 개행, 탭이동 문자만 들어올 경우 IllegalArgumentException 예외를 던진다.
+  - 이름에 `pobi`가 들어올 경우 `pobi`를 가진 이름이 생성되어야 한다.
+- [x] 이름들은 1 ~ 10개의 사이즈로 생성될 수 있다.
+  - 이름들 생성시 빈리스트가 들어올 경우 IllegalArgumentException 예외를 던진다.
+  - 이름들 생성시 사이즈가 11개인 리스트가 들어올 경우 IllegalArgumentException 예외를 던진다.
+  - 이름들 생성시 [aaa, bbb]가 들어올 경우 [aaa, bbb]를 가진 이름들이 생성되어야 한다.- [x] 이름들은 1 ~ 10개의 사이즈로 생성될 수 있다.
+- [x] 갬블러 생성시 카드를 가지고 있지 말아야 한다.
+  - 갬블러 생성시 카드를 가지고 있지 말아야 한다.
+  - 카드들의 점수 합이 21점 이상이면 겜블러는 더이상 카드를 뽑을 수 없다.
+  - 카드들의 점수 합이 1 ~ 20점 이내면 겜블러는 카드를 뽑을 수 있다.
+- [x] 갬블러들 1 ~ 10명으로 생성될 수 있다.
+  - 겜블러들 생성시 빈리스트가 들어올 경우 IllegalArgumentException 예외를 던진다.
+  - 겜블러들 생성시 사이즈가 11개인 리스트가 들어올 경우 IllegalArgumentException 예외를 던진다.
+  - 겜블러들 생성시 이름이 [aaa, bbb]가 들어올 경우 [aaa, bbb]를 가진 겜블러들이 생성되어야 한다.
+- [x] 카드들은 비어 있을 수 있으며 동일한 카드가 존재 할 수 없다.
+  - 카드들 생성시 [7클로버, 7클로버, 7클로버]를 입력할 경우 [7클로버]만 존재 해야 한다.
+  - 카드들 생성시 [3클로버, 9클로버, J클로버]를 입력할 경우 최소 값이 22가 나와야 한다.
+  - 카드들 생성시 [3클로버, 9클로버, J클로버]를 입력할 경우 21에 근접하며 최대한 넘지않도록 값을 뽑을 경우 22가 나와야 한다.
+  - 카드들 생성시 [3클로버, 9클로버, A클로버]를 입력할 경우 21에 근접하며 최대한 넘지않도록 값을 뽑을 경우 13이 나와야 한다.
+  - 카드들 생성시 [A스페이스, A하트, A클로버, A다이아]를 입력할 경우 21에 근접하며 최대한 넘지않도록 값을 뽑을 경우 14가 나와야 한다.
+- [x] 덱에 있는 카드를 모두 사용 했을 때 한장 더 뽑을 수 없다.
+  - 덱에 있는 카드를 모두 사용 했을 때 카드를 뽑을 경우 IllegalArgumentException 예외를 던진다.
+  - 덱에 52장의 카드가 있는 경우 52번의 카드를 뽑을 수 있다.
+- [x] 스코어 생성시 1 ~ 31의 숫자로만 생성될 수 있다.
+  - 스코어 생성시 0이 들어올 경우 IllegalArgumentException 예외를 던진다.
+  - 스코어 생성시 32가 들어올 경우 IllegalArgumentException 예외를 던진다.
+  - 스코어 생성시 1~31이 들어올 경우 들어온 숫자와 같은 스코어가 생성되어야 한다.
+
+### 프로그래밍 요구 사항
+- 모든 기능을 TDD로 구현해 단위 테스트가 존재해야 한다. 단, UI(System.out, System.in) 로직은 제외
+- indent(인덴트, 들여쓰기) depth를 2를 넘지 않도록 구현한다. 1까지만 허용한다.
+- 모든 엔티티를 작게 유지한다.
+- 3개 이상의 인스턴스 변수를 가진 클래스를 쓰지 않는다.
+- 딜러와 플레이어에서 발생하는 중복 코드를 제거해야 한다.
+- 함수(또는 메서드)의 길이가 15라인을 넘어가지 않도록 구현한다.
+- 기능을 구현하기 전에 README.md 파일에 구현할 기능 목록을 정리해 추가한다.
+- git의 commit 단위는 앞 단계에서 README.md 파일에 정리한 기능 목록 단위로 추가한다.
