+++ src/main/kotlin/blackjack/Main.kt
@@ -1,0 +1,7 @@
+package blackjack
+
+import blackjack.controller.BlackjackController
+
+fun main() {
+    BlackjackController.playBlackJack()
+}
