+++ src/main/kotlin/blackjack/domain/model/Card.kt
@@ -1,0 +1,8 @@
+package blackjack.domain.model
+
+data class Card(val pattern: Pattern, val sign: Sign) {
+
+    companion object {
+        fun of(pattern: Pattern, sign: Sign): Card = Card(pattern, sign)
+    }
+}
