+++ src/main/kotlin/blackjack/domain/model/Pattern.kt
@@ -1,0 +1,5 @@
+package blackjack.domain.model
+
+enum class Pattern {
+    SPACE, CLOVER, HEART, DIAMOND
+}
