+++ src/main/kotlin/blackjack/view/DrawState.kt
@@ -1,0 +1,7 @@
+package blackjack.view
+
+enum class DrawState {
+    HIT, STAY;
+
+    fun isHit() = this == HIT
+}
