--- src/main/kotlin/step2/blackjack/Main.kt
@@ -1,7 +1,0 @@
-package step2.blackjack
-
-import step2.blackjack.controller.BlackjackController
-
-fun main() {
-    BlackjackController.playBlackJack()
-}
