--- src/main/kotlin/step2/blackjack/domain/model/Card.kt
@@ -1,8 +1,0 @@
-package step2.blackjack.domain.model
-
-data class Card(val pattern: Pattern, val sign: Sign) {
-
-    companion object {
-        fun of(pattern: Pattern, sign: Sign): Card = Card(pattern, sign)
-    }
-}
