--- src/main/kotlin/step2/blackjack/domain/model/Deck.kt
@@ -1,20 +1,0 @@
-package step2.blackjack.domain.model
-
-@JvmInline
-value class Deck private constructor(private val cards: Cards) {
-
-    fun peek(): Card = cards.peek()
-
-    companion object {
-        fun all(): Deck {
-            val cards = Cards.from(Pattern.values()
-                .flatMap { pattern ->
-                    Sign.values()
-                        .map { sign -> Card.of(pattern, sign) }
-                }
-                .shuffled()
-                .toSet())
-            return Deck(cards)
-        }
-    }
-}
