--- src/main/kotlin/step2/blackjack/domain/model/Gambler.kt
@@ -1,15 +1,0 @@
-package step2.blackjack.domain.model
-
-data class Gambler(
-    val name: Name,
-    val cards: Cards
-) {
-
-    fun receive(card: Card) {
-        cards + card
-    }
-
-    companion object {
-        fun from(name: Name): Gambler = Gambler(name, Cards.empty())
-    }
-}
