--- src/main/kotlin/step2/blackjack/domain/model/Pattern.kt
@@ -1,5 +1,0 @@
-package step2.blackjack.domain.model
-
-enum class Pattern {
-    SPACE, CLOVER, HEART, DIAMOND
-}
