--- src/main/kotlin/step2/blackjack/view/DrawState.kt
@@ -1,7 +1,0 @@
-package step2.blackjack.view
-
-enum class DrawState {
-    HIT, STAY;
-
-    fun isHit() = this == HIT
-}
