+++ src/test/kotlin/study/model/Person.kt
@@ -1,0 +1,11 @@
+package study.model
+
+/**
+ * 사람 객체
+ * */
+data class Person(
+    val name: String,
+    val company: String,
+    val skills: List<Pair<String, String>>,
+    val languages: Map<String, Int>
+)
